unify the logic for integration and publicapi
