+++ server/internal/adapter/publicapi/schema_export.go
@@ -4,9 +4,10 @@ import (
 	"context"
 
 	"github.com/reearth/reearth-cms/server/internal/usecase/interfaces"
+	"github.com/reearth/reearth-cms/server/pkg/exporters"
+	"github.com/reearth/reearth-cms/server/pkg/id"
 	"github.com/reearth/reearth-cms/server/pkg/model"
 	"github.com/reearth/reearth-cms/server/pkg/schema"
-	"github.com/reearth/reearth-cms/server/pkg/value"
 	"github.com/reearth/reearthx/rerror"
 	"github.com/samber/lo"
 )
@@ -27,93 +28,35 @@ func (c *Controller) GetSchemaJSON(ctx context.Context, pKey, mKey string) (Sche
 		return SchemaJSON{}, rerror.ErrNotFound
 	}
 
-	return NewSchemaJSON(m.ID().String(), lo.ToPtr(m.Name()), lo.ToPtr(m.Description()), buildProperties(c.usecases, sp.Schema().Fields(), ctx)), nil
+	gsMap := buildGroupSchemaMap(ctx, sp.Schema(), c.usecases)
+	res := exporters.NewSchemaJSON(m.ID().String(), lo.ToPtr(m.Name()), lo.ToPtr(m.Description()), exporters.BuildProperties(ctx, sp.Schema().Fields(), gsMap))
+
+	return toSchemaJSON(res), nil
 }
 
-func buildProperties(uc *interfaces.Container, f schema.FieldList, ctx context.Context) map[string]interface{} {
-	properties := make(map[string]interface{})
-	for _, field := range f {
-		fieldType, format := determineTypeAndFormat(field.Type())
-		fieldSchema := map[string]interface{}{
-			"type":        fieldType,
-			"title":       field.Name(),
-			"description": field.Description(),
-		}
-		if format != "" {
-			fieldSchema["format"] = format
-		}
+func toSchemaJSON(s exporters.SchemaJSON) SchemaJSON {
+	return SchemaJSON{
+		Schema:      s.Schema,
+		Id:          s.Id,
+		Title:       s.Title,
+		Description: s.Description,
+		Type:        s.Type,
+		Properties:  s.Properties,
+	}
+}
 
+func buildGroupSchemaMap(ctx context.Context, sch *schema.Schema, uc *interfaces.Container) map[id.GroupID]*schema.Schema {
+	groupSchemaMap := make(map[id.GroupID]*schema.Schema)
+
+	for _, field := range sch.Fields() {
 		field.TypeProperty().Match(schema.TypePropertyMatch{
-			Text: func(f *schema.FieldText) {
-				if maxLength := f.MaxLength(); maxLength != nil {
-					fieldSchema["maxLength"] = *maxLength
-				}
-			},
-			TextArea: func(f *schema.FieldTextArea) {
-				if maxLength := f.MaxLength(); maxLength != nil {
-					fieldSchema["maxLength"] = *maxLength
-				}
-			},
-			RichText: func(f *schema.FieldRichText) {
-				if maxLength := f.MaxLength(); maxLength != nil {
-					fieldSchema["maxLength"] = *maxLength
-				}
-			},
-			Markdown: func(f *schema.FieldMarkdown) {
-				if maxLength := f.MaxLength(); maxLength != nil {
-					fieldSchema["maxLength"] = *maxLength
-				}
-			},
-			Integer: func(f *schema.FieldInteger) {
-				if min := f.Min(); min != nil {
-					fieldSchema["minimum"] = *min
-				}
-				if max := f.Max(); max != nil {
-					fieldSchema["maximum"] = *max
-				}
-			},
-			Number: func(f *schema.FieldNumber) {
-				if min := f.Min(); min != nil {
-					fieldSchema["minimum"] = *min
-				}
-				if max := f.Max(); max != nil {
-					fieldSchema["maximum"] = *max
-				}
-			},
-			Group: func(f *schema.FieldGroup) {
-				gs, _ := uc.Schema.FindByGroup(ctx, f.Group(), nil)
-				if gs != nil {
-					fieldSchema["items"] = buildProperties(uc, gs.Fields(), ctx)
+			Group: func(fg *schema.FieldGroup) {
+				groupSchema, err := uc.Schema.FindByGroup(ctx, fg.Group(), nil)
+				if err == nil {
+					groupSchemaMap[fg.Group()] = groupSchema
 				}
 			},
 		})
-
-		properties[field.Key().String()] = fieldSchema
 	}
-	return properties
-}
-
-func determineTypeAndFormat(t value.Type) (string, string) {
-	switch t {
-	case value.TypeText, value.TypeTextArea, value.TypeRichText, value.TypeMarkdown, value.TypeSelect, value.TypeTag, value.TypeReference:
-		return "string", ""
-	case value.TypeInteger:
-		return "integer", ""
-	case value.TypeNumber:
-		return "number", ""
-	case value.TypeBool, value.TypeCheckbox:
-		return "boolean", ""
-	case value.TypeDateTime:
-		return "string", "date-time"
-	case value.TypeURL:
-		return "string", "uri"
-	case value.TypeAsset:
-		return "string", "binary"
-	case value.TypeGroup:
-		return "array", ""
-	case value.TypeGeometryObject, value.TypeGeometryEditor:
-		return "object", ""
-	default:
-		return "string", ""
-	}
+	return groupSchemaMap
 }
