+++ server/pkg/exporters/schema_json_test.go
@@ -1,12 +1,9 @@
-package publicapi
+package exporters
 
 import (
 	"context"
 	"testing"
 
-	"github.com/reearth/reearth-cms/server/internal/infrastructure/memory"
-	"github.com/reearth/reearth-cms/server/internal/usecase/interactor"
-	"github.com/reearth/reearth-cms/server/internal/usecase/interfaces"
 	"github.com/reearth/reearth-cms/server/pkg/group"
 	"github.com/reearth/reearth-cms/server/pkg/id"
 	"github.com/reearth/reearth-cms/server/pkg/schema"
@@ -18,8 +15,6 @@ import (
 
 func TestBuildProperties(t *testing.T) {
 	ctx := context.Background()
-	r := memory.New()
-	uc := &interfaces.Container{Schema: interactor.NewSchema(r, nil)}
 
 	wid := accountdomain.NewWorkspaceID()
 	pid := id.NewProjectID()
@@ -43,13 +38,11 @@ func TestBuildProperties(t *testing.T) {
 
 	// group schema
 	gs := schema.New().ID(id.NewSchemaID()).Workspace(wid).Project(pid).Fields([]*schema.Field{gsf}).MustBuild()
-	lo.Must0(r.Schema.Save(ctx, gs))
 
 	// group
 	gid := id.NewGroupID()
 	gkey := id.RandomKey()
 	g := group.New().ID(gid).Name("group").Project(pid).Key(gkey).Schema(gs.ID()).MustBuild()
-	lo.Must0(r.Group.Save(ctx, g))
 
 	// group field
 	fId3 := id.NewFieldID()
@@ -72,8 +65,7 @@ func TestBuildProperties(t *testing.T) {
 	sf6 := schema.NewField(schema.NewURL().TypeProperty()).ID(fId6).Key(sfKey6).MustBuild()
 
 	fieldList := schema.FieldList{sf1, sf2, sf3, sf4, sf5, sf6}
-	s1 := schema.New().NewID().Workspace(wid).Project(pid).Fields(schema.FieldList{sf2}).MustBuild()
-	lo.Must0(r.Schema.Save(ctx, s1))
+	gsMap :=  map[id.GroupID]*schema.Schema{gid: gs}
 
 	expectedProperties := map[string]interface{}{
 		sfKey1.String(): map[string]interface{}{
@@ -121,7 +113,9 @@ func TestBuildProperties(t *testing.T) {
 		},
 	}
 
-	properties := buildProperties(uc, fieldList, ctx)
+
+
+	properties := BuildProperties(ctx, fieldList, gsMap)
 	assert.Equal(t, expectedProperties, properties)
 }
